requirements.txt
requests==2.32.3
python-metar==1.4.0
python-dotenv==1.0.1